pyproject.toml
[tool.poetry]
name = "recomendacaoprodutos"
version = "0.1.0"
description = ""
authors = ["Amarildo <amarildozj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
scikit-learn = "^1.5.2"
pandas = "^2.2.3"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
sqlalchemy = "^2.0.36"
pyodbc = "^5.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
